pyproject.toml
[project]
name = "pg-spec-extractor"
version = "1.0.0"
description = "Paraglider spec extraction pipeline — crawl, extract, normalize, store"
requires-python = ">=3.11"
dependencies = [
    "crawl4ai>=0.8.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "httpx>=0.28",
    "ollama>=0.4",
    "python-dotenv>=1.0",
    "typer>=0.12",
]

[project.scripts]
pg-extract = "src.pipeline:app"
